+++ Controllers/HomeController.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using EmployeeApp.Models;
 
+
 namespace EmployeeApp.Controllers;
 
 public class HomeController : Controller
