+++ Models/Company.cs
@@ -1,15 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace EmployeeApp;
+namespace EmployeeApp.Models;
 
-public partial class Employee
+public partial class Company
 {
     public int Id { get; set; }
 
-    public string EmployeeName { get; set; } = null!;
-
-    public string PasswordHash { get; set; } = null!;
+    public string CompanyName { get; set; } = null!;
 
     public DateTime? CreatedAt { get; set; }
 
@@ -18,6 +16,4 @@ public partial class Employee
     public virtual ICollection<Department> Departments { get; set; } = new List<Department>();
 
     public virtual ICollection<Division> Divisions { get; set; } = new List<Division>();
-
-    public virtual ICollection<License> Licenses { get; set; } = new List<License>();
 }
