+++ Models/Division.cs
@@ -1,23 +1,25 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace EmployeeApp;
+namespace EmployeeApp.Models;
 
-public partial class Department
+public partial class Division
 {
     public int Id { get; set; }
 
     public int CompanyId { get; set; }
 
-    public string DapartmentName { get; set; } = null!;
+    public int? DepartmentId { get; set; }
+
+    public string DivisionName { get; set; } = null!;
 
     public DateTime? CreatedAt { get; set; }
 
-    public DateTime? UpdatedAt { get; set; }
+    public DateTime? UpdateAt { get; set; }
 
     public virtual Company Company { get; set; } = null!;
 
-    public virtual ICollection<Division> Divisions { get; set; } = new List<Division>();
+    public virtual Department? Department { get; set; }
 
     public virtual ICollection<Employee> Employees { get; set; } = new List<Employee>();
 }
